Do item spell id substitutions in core.
Previously we replaced the spell id used by items that had the data for the spell itself changed in item_template as if it were a change to the item. That meant we actually had wrong item data for those patches, but it made the effects look correct in the 1.12 client. If you were to play with the actual patch appropriate client however, all of those items would show wrong effects. Doing spell id substitutions for newer clients in the core is a better way to handle this, as it means we can have 100% correct and patch appropriate data for items in the database. Also this way there is no way of missing an item that needs a replacement spell id for lower content patches, since all the items are checked on startup.