Rewrite "UPDATETYPE_CREATE_OBJECT2" usage
This value seem to be used only when an object is realy created by the server.
When an object already exist in the map we have to send "UPDATETYPE_CREATE_OBJECT".
This will have huge impact with all the animation at effective spawn.
Some new problem may appear after this.
If this is confirmed we can rename "UPDATETYPE_CREATE_OBJECT2" to
something more correct.