Create item instances with guild charter flag from template
Preserves ITEM_FLAG_CHARTER from the item template when creating an
instance of the item. The 1.6 client relies on this flag to correctly
recognize the item as a guild charter. This both enables the player to
register the charter and causes the guild info (name and master) to
appear in the charter's tooltip where it is otherwise missing.