Add more accurate script for Eye of Naxxramas.
Closes https://github.com/slowtorta/turtlewow-bug-tracker/issues/1558

Went into Stratholme solo in Classic with a rogue to research exactly how this npc behaves. It starts casting the spell to summon the Gargoyles immediatelly on aggro, which has a 10 second cast time, so that's why it looks like the mob doesn't attack, but it actually does once it finishes the cast. And the Gargoyles are guardian pets, which continue to follow the stealthed Eye of Naxxramas if you wipe. They don't despawn on evade. Also there is no sound being played on aggro like in the old script.