Fix cooldown of Shadowguard.
The table stores cooldown in seconds, but somebody thought it was milliseconds and gave it a 3500 second cooldown. In vmangos the cooldown system was reworked to handle cooldowns that are less than a second, so its stored in milliseconds there, but not in Turtle core, since that commit has not been applied here.

Fixes https://github.com/slowtorta/turtlewow-bug-tracker/issues/1550