Use different anti abuse mechanic for In Dreams.
RMJ requested to remove the teleport to Isillien. Now spawning guards into the tower instead, to prevent players from going there. Pathing is broken there, and the quest will break if player hides inside the tower. Isillien must be able to aggro the player for next part of event to be triggered.