Various DB fixes from.
Fixed a typo in the quest Tower of Lapidis II.
Assigned a dungeon type to the quest The Stockade's Search.
Fixed grammar and style in the quest Arugal's Folly.
Fixed a typo in the quest The Old Church of Westfall X.
Fixed typo in the quest Tesala's Word.
Removed description from the trainer's spell Savage Bite.
Fixed a typo in the mushroom's growing stage.
Fixed a typo in the quest A Friend of the Tribe.
Fixed the misleading description of the spell Moonkin Form.
Fixed a typo in the item Tyrant's Crown.
Fixed a typo in the quest The Depths of Karazhan VII.
Fixed incomplete progress text for the quest Special Forces.
Fixed a typo in the quest Da Unforgiving Lands.
Fixed a typo in the quest Goblin Engineering At Its Finest!
Fixed a typo in quest Willing to Serve.