Cap level based spell miss chance.
Fixes https://github.com/slowtorta/turtlewow-bug-tracker/issues/3407

It becomes mathematically impossible to hit a target with a spell if level difference is more than 13, yet on the official classic servers you can still hit people with spells, so there's probably supposed to be some kind of minimum. A lower cap of 33% is equivalent to 10 levels difference max.