It shouldn't be possible to bubble or stealth with WSG flag.
In classic you get an error message and it doesn't let you cast it, instead of casting it and causing flag to drop. The attribute that I believe controls this was only added to immunity spells in patch 1.9, so that would indicate it was possible to bop people to make them drop flag in early vanilla.