Fix disproportionate rep rewards for 30-39 WSG quests
Every other first-time turn-in of Warsong Gulch Mark of Honor rewards 500 reputation for Silverwing / Warsong, except for the 30-39 tier.  This makes for a peculiar "meta" where the best time to turn in this quest is between these levels, and it defies convention.

Instead of 4x 200 reputation, these quests now give 1x 500 reputation like all other iterations of the quests.