Fix engineering pet spells exploit.
This was fixed in vmangos back in 2018. Someone must have replaced the server side skill_line_ability data with the contents of the dbc from client, thus reintroducing the exploit. The dbc was made into a db table specifically to fix this issue, so we can alter the data in vmangos.